refactor(core): move relative positional encoding primitives to shared nn layer
build_relative_positional_encoding (pure sinusoidal Transformer-XL table, no
cohere-private state) lived in models/cohere/encoder_graph.rs but was imported
by firered_aed, parakeet_ctc, and parakeet_tdt, leaking CohereTranscribeEncoderError
into three unrelated families. Move it to nn/encoder.rs alongside conformer_block
(the shared IR-boundary module it already documented a dependency on) and make
it generic over the caller's error type via an overflow_err closure, matching
every other nn/ builder's convention. Each family now supplies its own error
variant at the call site instead of mapping cohere's.

Pure move + signature neutralization: no arithmetic changed. Verified with real
installed .oasr packs (cohere-transcribe-03-2026, parakeet-ctc-0.6b,
parakeet-tdt-0.6b-v3, firered-aed-l-v2) that end-to-end transcription output is
byte-identical before and after this change.
